Change IRI and add fixes.
For now I gave the ontology a routable name -- even if a stupid one -- because I have been having trouble with loading from the catalog.
Also split the `hasVendor` property into `hasVendor` (for catalog entry) and `equipmentVendor` (for `LabEquipment`).  That seems to sit better with the OWL DL syntax restrictions.  I should confess that those are a little mysterious to me, so this may not have been required.