Commit branch-specific TTL files.
This is an ugliness made necessary by the fact that the files must
have their own URL in them.  Also these files must be available from
GitHub until we find a permanent home for them.

So even though these are generated by a compilation process, they need
to be committed.